Revert wrong package for postgresql
