+++ parser/sql/dialect/postgresql/src/main/resources/META-INF/services/org.apache.shardingsphere.sql.parser.spi.DialectSQLParserFacade
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.sql.parser.presto.parser.PostgreSQLParserFacade
+org.apache.shardingsphere.sql.parser.postgresql.parser.PostgreSQLParserFacade
